+++ src/assets.d.ts
@@ -1,0 +1,4 @@
+declare module '*.jpg';
+declare module '*.png';
+declare module '*.jpeg';
+declare module '*.gif';
