+++ src/screens/home/components/Header/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import {Container, Title, ProfileImage} from './styles';
+
+const Header = () => {
+  return (
+    <Container>
+      <Title>CryptoTracker Pro</Title>
+      <ProfileImage source={require('../../../../../assets/img/user.png')} />
+    </Container>
+  );
+};
+
+export default Header;
