+++ src/styles/theme.ts
@@ -1,9 +1,11 @@
-const THEME = {
+const theme = {
   blue: '#385775',
-  ligthBlue: '#7778a1',
-  priceGreen: '#0A8150',
+  addText: '#395675',
+  priceGreen: '#0b8050',
   priceRed: '#DE3617',
   divider: '#B7C0C6',
+  white: '#FFFFFF',
+  priceText: '#212b36',
 };
 
-export default THEME;
+export default theme;
